Changed: Initialized Clean Architecture configuration.
