+++ app/src/main/java/io/jcasas/weatherdagger2example/data/source/external/WeatherService.kt
@@ -16,8 +16,8 @@
 
 package io.jcasas.weatherdagger2example.data.source.external
 
-import io.jcasas.weatherdagger2example.data.source.model.ForecastResponse
-import io.jcasas.weatherdagger2example.data.source.model.WeatherResponse
+import com.juancasasm.android.weatherexample.domain.ForecastResponse
+import com.juancasasm.android.weatherexample.domain.WeatherResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 import retrofit2.Call
