--- app/src/main/java/io/jcasas/weatherdagger2example/data/source/model/ForecastResponse.kt
@@ -1,6 +1,0 @@
-package io.jcasas.weatherdagger2example.data.source.model
-
-/**
- * Created by jcasas on 10/6/17.
- */
-class ForecastResponse(val list: ArrayList<Forecast>)
