+++ core/src/main/java/com/juancasasm/android/weatherexample/data/weather/WeatherDataSource.kt
@@ -1,0 +1,10 @@
+package com.juancasasm.android.weatherexample.data.weather
+
+import com.juancasasm.android.weatherexample.domain.Coordinates
+import com.juancasasm.android.weatherexample.domain.Weather
+
+interface WeatherDataSource {
+
+    suspend fun getCurrent(coordinates: Coordinates): Weather
+
+}
