+++ core/src/main/java/com/juancasasm/android/weatherexample/domain/ForecastResponse.kt
@@ -1,0 +1,8 @@
+package com.juancasasm.android.weatherexample.domain
+
+import com.juancasasm.android.weatherexample.domain.Forecast
+
+/**
+ * Created by jcasas on 10/6/17.
+ */
+class ForecastResponse(val list: ArrayList<Forecast>)
