+++ core/src/main/java/com/juancasasm/android/weatherexample/domain/Weather.kt
@@ -14,16 +14,10 @@
  * limitations under the License.
  */
 
-package io.jcasas.weatherdagger2example.data.source.model
-
-import java.util.*
+package com.juancasasm.android.weatherexample.domain
 
 /**
  * Created by jcasas on 8/10/17.
  */
-class WeatherResponse(val weather:LinkedList<Weather>,
-                      val coord:Coord,
-                      val id:Integer,
-                      val name:String,
-                      val cod:Integer,
-                      val main:Conditions)
+// TODO Make it to contain all we need.
+class Weather(val id: Int, val main: String, val description: String)
